// src/generated/model/bannerCreateModel.ts
/**
 * Datafood Server API
 * API contract of Datafood server application BRASIDAS
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BannerCreateModel { 
    /**
     * Banner message
     */
    message: string;
    /**
     * Banner type
     */
    type: string;
    isActive: boolean;
    /**
     * Date de création du document
     */
    dateCreation: string;
}